refactor: rename types.ts
