+++ src/modules/i18n.ts
@@ -1,5 +1,5 @@
 import { createI18n } from 'vue-i18n'
-import type { UserModule } from '@/types'
+import type { UserModule } from '~/src/types'
 
 // Import i18n resources
 // https://vitejs.dev/guide/features.html#glob-import
