+++ src/modules/nprogress.ts
@@ -1,5 +1,5 @@
 import NProgress from 'nprogress'
-import type { UserModule } from '@/types'
+import type { UserModule } from '~/src/types'
 
 export const install: UserModule = ({ isClient, router }) => {
   if (isClient) {
