+++ src/modules/pinia.ts
@@ -1,5 +1,5 @@
 import { createPinia } from 'pinia'
-import type { UserModule } from '@/types'
+import type { UserModule } from '~/src/types'
 
 // Setup Pinia
 // https://pinia.esm.dev/
