+++ src/modules/pwa.ts
@@ -1,4 +1,4 @@
-import type { UserModule } from '@/types'
+import type { UserModule } from '~/src/types'
 
 // https://github.com/antfu/vite-plugin-pwa#automatic-reload-when-new-content-available
 export const install: UserModule = ({ isClient, router }) => {
